Remove the last ValidatedModelForm and thus BsForm
