+++ app/settings/tokens/-components/edit-form/styles.scss
@@ -1,0 +1,11 @@
+.Scope {
+    display: flex;
+
+    input {
+        flex-shrink: 0;
+    }
+}
+
+.Scope__description {
+    font-weight: normal;
+}
