--- lib/osf-components/addon/components/validated-model-form/template.hbs
@@ -1,78 +1,0 @@
-<BsForm
-    @onSubmit={{perform this.saveModelTask}}
->
-    {{yield (hash
-        model=this.model
-        changeset=this.changeset
-        disabled=this.inputsDisabled
-        submit=(perform this.saveModelTask)
-        shouldShowMessages=this.shouldShowMessages
-        rollback=(action this.rollback)
-        isDirty=this.changeset.isDirty
-
-        checkbox=(
-            component
-            'validated-input/checkbox'
-            changeset=this.changeset
-            model=this.model
-            shouldShowMessages=this.shouldShowMessages
-            disabled=this.inputsDisabled
-        )
-        checkboxes=(
-            component
-            'validated-input/checkboxes'
-            changeset=this.changeset
-            model=this.model
-            shouldShowMessages=this.shouldShowMessages
-            disabled=this.inputsDisabled
-        )
-        date=(
-            component
-            'validated-input/date'
-            changeset=this.changeset
-            model=this.model
-            shouldShowMessages=this.shouldShowMessages
-            disabled=this.inputsDisabled
-        )
-        recaptcha=(
-            component
-            'validated-input/recaptcha'
-            changeset=this.changeset
-            model=this.model
-            shouldShowMessages=this.shouldShowMessages
-            disabled=this.inputsDisabled
-        )
-        text=(
-            component
-            'validated-input/text'
-            changeset=this.changeset
-            model=this.model
-            shouldShowMessages=this.shouldShowMessages
-            disabled=this.inputsDisabled
-        )
-        textarea=(
-            component
-            'validated-input/textarea'
-            changeset=this.changeset
-            model=this.model
-            shouldShowMessages=this.shouldShowMessages
-            disabled=this.inputsDisabled
-        )
-        select=(
-            component
-            'validated-input/power-select'
-            changeset=this.changeset
-            model=this.model
-            shouldShowMessages=this.shouldShowMessages
-            disabled=this.inputsDisabled
-        )
-        custom=(
-            component
-            'validated-input/custom'
-            changeset=this.changeset
-            model=this.model
-            shouldShowMessages=this.shouldShowMessages
-            disabled=this.inputDisabled
-        )
-    )}}
-</BsForm>
