--- lib/osf-components/app/components/validated-model-form/component.js
@@ -1,1 +1,0 @@
-export { default } from 'osf-components/components/validated-model-form/component';
